Add three-level documentation navigation tree with grouped subsections
Introduces `doc.intro_group()` for 3-level tree nesting in the sidebar.
Each section now uses subsection modules to group related elements
(e.g., Controls > Buttons > Button, Button Group, etc.), improving
discoverability while preserving the curated semantic ordering.
