+++ website/documentation/content/doc/__init__.py
@@ -4,6 +4,7 @@ from .api import (
     extra_column,
     get_page,
     intro,
+    intro_group,
     part,
     redirects,
     reference,
@@ -19,6 +20,7 @@ __all__ = [
     'extra_column',
     'get_page',
     'intro',
+    'intro_group',
     'part',
     'redirects',
     'reference',
