+++ website/documentation/content/subsection_buttons.py
@@ -1,0 +1,16 @@
+from . import (
+    badge_documentation,
+    button_documentation,
+    button_dropdown_documentation,
+    button_group_documentation,
+    doc,
+    fab_documentation,
+)
+
+doc.title('Buttons')
+
+doc.intro(button_documentation)
+doc.intro(button_group_documentation)
+doc.intro(button_dropdown_documentation)
+doc.intro(fab_documentation)
+doc.intro(badge_documentation)
