+++ website/documentation/content/subsection_chips.py
@@ -1,0 +1,10 @@
+from . import (
+    chip_documentation,
+    doc,
+    input_chips_documentation,
+)
+
+doc.title('Chips')
+
+doc.intro(chip_documentation)
+doc.intro(input_chips_documentation)
