+++ website/documentation/content/subsection_choices.py
@@ -1,0 +1,16 @@
+from . import (
+    checkbox_documentation,
+    doc,
+    radio_documentation,
+    select_documentation,
+    switch_documentation,
+    toggle_documentation,
+)
+
+doc.title('Choices')
+
+doc.intro(checkbox_documentation)
+doc.intro(radio_documentation)
+doc.intro(select_documentation)
+doc.intro(switch_documentation)
+doc.intro(toggle_documentation)
