+++ website/documentation/content/subsection_client_storage.py
@@ -1,0 +1,14 @@
+from . import (
+    clipboard_documentation,
+    doc,
+    event_documentation,
+    run_javascript_documentation,
+    storage_documentation,
+)
+
+doc.title('Client & Storage')
+
+doc.intro(run_javascript_documentation)
+doc.intro(clipboard_documentation)
+doc.intro(event_documentation)
+doc.intro(storage_documentation)
