+++ website/documentation/content/subsection_diagrams_messages.py
@@ -1,0 +1,10 @@
+from . import (
+    chat_message_documentation,
+    doc,
+    mermaid_documentation,
+)
+
+doc.title('Diagrams & Messages')
+
+doc.intro(chat_message_documentation)
+doc.intro(mermaid_documentation)
