+++ website/documentation/content/subsection_editors_data.py
@@ -1,0 +1,14 @@
+from . import (
+    code_documentation,
+    doc,
+    editor_documentation,
+    json_editor_documentation,
+    log_documentation,
+)
+
+doc.title('Editors & Structured Data')
+
+doc.intro(editor_documentation)
+doc.intro(code_documentation)
+doc.intro(json_editor_documentation)
+doc.intro(log_documentation)
