+++ website/documentation/content/subsection_inputs.py
@@ -1,0 +1,20 @@
+from . import (
+    color_input_documentation,
+    date_input_documentation,
+    doc,
+    input_documentation,
+    number_documentation,
+    textarea_documentation,
+    time_input_documentation,
+    upload_documentation,
+)
+
+doc.title('Inputs')
+
+doc.intro(input_documentation)
+doc.intro(textarea_documentation)
+doc.intro(number_documentation)
+doc.intro(color_input_documentation)
+doc.intro(date_input_documentation)
+doc.intro(time_input_documentation)
+doc.intro(upload_documentation)
