+++ website/documentation/content/subsection_media.py
@@ -1,0 +1,10 @@
+from . import (
+    audio_documentation,
+    doc,
+    video_documentation,
+)
+
+doc.title('Media')
+
+doc.intro(audio_documentation)
+doc.intro(video_documentation)
