+++ website/documentation/content/subsection_navigation.py
@@ -1,0 +1,94 @@
+from nicegui import app, ui
+
+from . import (
+    doc,
+    download_documentation,
+    navigate_documentation,
+    page_title_documentation,
+)
+
+doc.title('Navigation')
+
+doc.intro(page_title_documentation)
+doc.intro(navigate_documentation)
+doc.intro(download_documentation)
+
+doc.redirects['open'] = 'navigate#ui_navigate_to_(formerly_ui_open)'
+doc.text('ui.open', f'''
+    The `ui.open` function has been removed.
+    Use [`ui.navigate.to`]({doc.redirects["open"]}) instead.
+''')
+
+
+@doc.demo(app.add_static_files)
+def add_static_files_demo():
+    from nicegui import app
+
+    app.add_static_files('/examples', 'examples')
+    ui.label('Some NiceGUI Examples').classes('text-h5')
+    ui.link('AI interface', '/examples/ai_interface/main.py')
+    ui.link('Custom FastAPI app', '/examples/fastapi/main.py')
+    ui.link('Authentication', '/examples/authentication/main.py')
+
+
+@doc.demo(app.add_media_files)
+def add_media_files_demo():
+    from pathlib import Path
+
+    import httpx
+
+    from nicegui import app
+
+    media = Path('media')
+    # media.mkdir(exist_ok=True)
+    # r = httpx.get('https://cdn.coverr.co/videos/coverr-cloudy-sky-2765/1080p.mp4')
+    # (media  / 'clouds.mp4').write_bytes(r.content)
+    # app.add_media_files('/my_videos', media)
+    # ui.video('/my_videos/clouds.mp4')
+    # END OF DEMO
+    ui.video('https://cdn.coverr.co/videos/coverr-cloudy-sky-2765/1080p.mp4')
+
+
+@doc.demo('Add HTML to the page', '''
+    You can add HTML to the page by calling `ui.add_head_html` or `ui.add_body_html`.
+    This is useful for adding custom CSS styles or JavaScript code.
+''')
+def add_head_html_demo():
+    ui.add_head_html('''
+        <style>
+            .my-red-label {
+                color: Crimson;
+                font-weight: bold;
+            }
+        </style>
+    ''')
+    ui.label('RED').classes('my-red-label')
+
+
+@doc.auto_execute
+@doc.demo('API Responses', '''
+    NiceGUI is based on [FastAPI](https://fastapi.tiangolo.com/).
+    This means you can use all of FastAPI's features.
+    For example, you can implement a RESTful API in addition to your graphical user interface.
+    You simply import the `app` object from `nicegui`.
+    Or you can run NiceGUI on top of your own FastAPI app by using `ui.run_with(app)` instead of starting a server automatically with `ui.run()`.
+
+    You can also return any other FastAPI response object inside a page function.
+    For example, you can return a `RedirectResponse` to redirect the user to another page if certain conditions are met.
+    This is used in our [authentication demo](https://github.com/zauberzeug/nicegui/tree/main/examples/authentication/main.py).
+''')
+def fastapi_demo():
+    import random
+
+    from nicegui import app
+
+    @app.get('/random/{max}')
+    def generate_random_number(max: int):
+        return {'min': 0, 'max': max, 'value': random.randint(0, max)}
+
+    # @ui.page('/')
+    def page():
+        max = ui.number('max', value=100)
+        ui.button('generate random number',
+                  on_click=lambda: ui.navigate.to(f'/random/{max.value:.0f}'))
+    page()  # HIDE
