+++ website/documentation/content/subsection_pickers.py
@@ -1,0 +1,12 @@
+from . import (
+    color_picker_documentation,
+    date_documentation,
+    doc,
+    time_documentation,
+)
+
+doc.title('Pickers')
+
+doc.intro(color_picker_documentation)
+doc.intro(date_documentation)
+doc.intro(time_documentation)
