+++ website/documentation/content/subsection_spatial_trees.py
@@ -1,0 +1,12 @@
+from . import (
+    doc,
+    leaflet_documentation,
+    scene_documentation,
+    tree_documentation,
+)
+
+doc.title('Spatial & Trees')
+
+doc.intro(scene_documentation)
+doc.intro(leaflet_documentation)
+doc.intro(tree_documentation)
