+++ website/documentation/content/subsection_special_controls.py
@@ -1,0 +1,10 @@
+from . import (
+    codemirror_documentation,
+    doc,
+    xterm_documentation,
+)
+
+doc.title('Special')
+
+doc.intro(codemirror_documentation)
+doc.intro(xterm_documentation)
