+++ website/documentation/content/subsection_tables_grids.py
@@ -1,0 +1,10 @@
+from . import (
+    aggrid_documentation,
+    doc,
+    table_documentation,
+)
+
+doc.title('Tables')
+
+doc.intro(table_documentation)
+doc.intro(aggrid_documentation)
