+++ website/documentation/content/subsection_test_fixtures.py
@@ -1,0 +1,10 @@
+from . import (
+    doc,
+    screen_documentation,
+    user_documentation,
+)
+
+doc.title('Test Fixtures')
+
+doc.intro(user_documentation)
+doc.intro(screen_documentation)
