+++ website/documentation/content/subsection_text_markup.py
@@ -1,0 +1,14 @@
+from . import (
+    doc,
+    label_documentation,
+    link_documentation,
+    markdown_documentation,
+    restructured_text_documentation,
+)
+
+doc.title('Text & Markup')
+
+doc.intro(label_documentation)
+doc.intro(link_documentation)
+doc.intro(markdown_documentation)
+doc.intro(restructured_text_documentation)
