+++ website/documentation/content/subsection_timers_input.py
@@ -1,0 +1,10 @@
+from . import (
+    doc,
+    keyboard_documentation,
+    timer_documentation,
+)
+
+doc.title('Timers & Input')
+
+doc.intro(timer_documentation)
+doc.intro(keyboard_documentation)
