+++ website/documentation/content/subsection_value_interaction.py
@@ -1,0 +1,16 @@
+from . import (
+    doc,
+    joystick_documentation,
+    knob_documentation,
+    range_documentation,
+    rating_documentation,
+    slider_documentation,
+)
+
+doc.title('Value & Interaction')
+
+doc.intro(slider_documentation)
+doc.intro(range_documentation)
+doc.intro(knob_documentation)
+doc.intro(rating_documentation)
+doc.intro(joystick_documentation)
